Fix bug with -m.
The -m option would try to download a cetk from the CDN even if -c is
missing and exit the program if doing so failed. However, failure to
download a cetk is not a failure for -m (unlike -c, where the user
requires the decryption key).

This commit also prevents writing zero-length meta files (tmd/cetk).